get vessel tracks worker working
