+++ libs/api-types/src/index.ts
@@ -5,15 +5,16 @@ export * from './downloadActivity'
 export * from './events'
 export * from './geometries'
 export * from './i18n'
+export * from './identity-vessel'
 export * from './pagination'
-export * from './vessel-report'
 export * from './reports'
 export * from './resources'
+export * from './tracks'
 export * from './stats'
 export * from './user-applications'
 export * from './user'
-export * from './vessel'
-export * from './identity-vessel'
-export * from './vesselGroups'
 export * from './vessel-insights'
+export * from './vessel-report'
+export * from './vessel'
+export * from './vesselGroups'
 export * from './workspaces'
