+++ libs/deck-layer-composer/src/resolvers/fourwings.ts
@@ -110,7 +110,8 @@ export const resolveDeckFourwingsLayerProps = (
     hoveredFeatures: interactions,
     minVisibleValue: dataview.config?.minVisibleValue,
     maxVisibleValue: dataview.config?.maxVisibleValue,
-    debug: dataview.config?.debug ?? false,
+    // debug: dataview.config?.debug ?? false,
+    debug: true,
     visible: dataview.config?.visible ?? true,
     colorRampWhiteEnd: dataview.config?.colorRampWhiteEnd ?? false,
     ...(tilesUrl && { tilesUrl }),
