+++ libs/deck-loaders/project.json
@@ -18,17 +18,48 @@
       }
     },
     "dist:workers": {
+      "executor": "nx:run-commands",
+      "options": {
+        "commands": [
+          "nx fourwings:worker deck-loaders",
+          "nx tracks:worker deck-loaders",
+          "rm -rf dist/libs/deck-loaders/workers/package.json"
+        ],
+        "parallel": false
+      }
+    },
+    "fourwings:worker": {
       "executor": "@nx/esbuild:esbuild",
       "outputs": ["{options.outputPath}"],
       "options": {
         "outputPath": "dist/libs/deck-loaders/workers",
         "outputFileName": "fourwings-worker.js",
         "main": "libs/deck-loaders/src/fourwings/workers/fourwings-worker.ts",
+        "generateExportsField": true,
         "tsConfig": "libs/deck-loaders/tsconfig.lib.json",
         "watch": false,
         "target": "esnext",
         "platform": "browser",
-        "deleteOutputPath": true,
+        "deleteOutputPath": false,
+        "bundle": true,
+        "thirdParty": true,
+        "generatePackageJson": false,
+        "minify": false
+      }
+    },
+    "tracks:worker": {
+      "executor": "@nx/esbuild:esbuild",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/deck-loaders/workers",
+        "outputFileName": "vessel-tracks-worker.js",
+        "main": "libs/deck-loaders/src/vessels/workers/tracks-worker.ts",
+        "generateExportsField": true,
+        "tsConfig": "libs/deck-loaders/tsconfig.lib.json",
+        "watch": false,
+        "target": "esnext",
+        "platform": "browser",
+        "deleteOutputPath": false,
         "bundle": true,
         "thirdParty": true,
         "generatePackageJson": false,
