+++ libs/deck-loaders/src/fourwings/lib/types.ts
@@ -36,6 +36,7 @@ export type ParseFourwingsOptions = {
   noDataValue?: number
   sublayers: number
   buffersLength: number[]
+  workerUrl?: string
 }
 
 export type FourwingsLoaderOptions = LoaderOptions & {
