+++ libs/deck-loaders/src/loaders.config.ts
@@ -1,0 +1,6 @@
+export const IS_PRODUCTION =
+  typeof process !== 'undefined' ? process?.env?.NODE_ENV === 'production' : false
+export const PATH_BASENAME =
+  typeof process !== 'undefined'
+    ? process.env?.NEXT_PUBLIC_URL || (IS_PRODUCTION ? '/map' : '')
+    : ''
