+++ libs/deck-loaders/src/vessels/tracks-loader.ts
@@ -1,5 +1,6 @@
 import type { Loader, LoaderWithParser } from '@loaders.gl/loader-utils'
 import packageJson from '../../package.json'
+import { PATH_BASENAME } from '../loaders.config'
 import { parseTrack } from './lib/parse-tracks'
 
 /**
@@ -15,7 +16,11 @@ export const VesselTrackWorkerLoader: Loader = {
   extensions: ['pbf'],
   mimeTypes: ['application/x-protobuf', 'application/octet-stream', 'application/protobuf'],
   worker: true,
-  options: {},
+  options: {
+    'vessel-tracks': {
+      workerUrl: `${PATH_BASENAME}/workers/vessel-tracks-worker.js`,
+    },
+  },
 }
 
 /**
@@ -25,6 +30,5 @@ export const VesselTrackLoader: LoaderWithParser = {
   ...VesselTrackWorkerLoader,
   parse: async (arrayBuffer) => parseTrack(arrayBuffer),
   parseSync: parseTrack,
-  worker: false,
   binary: true,
 }
